// ui/src/lib/api/generated/model/invite.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire API
 * Centoire community platform API
 * OpenAPI spec version: 0.1.0
 */
import type { InviteRole } from './inviteRole';
import type { InviteStatus } from './inviteStatus';

export interface Invite {
  id: string;
  email: string;
  role: InviteRole;
  status: InviteStatus;
  expiresAt: string;
  /** @nullable */
  acceptedAt: string | null;
  createdAt: string;
}